combine layout stuff in one file
